feat: implement different states of navbar
Works pretty fine, but I encountered a small issue with the dropdown menu.
If it's clicked, you come to the section you wanted to, but the screen freezes until you click somewhere else. After a click, you come back to the hero section.
Because of this stuff, I commented the dropdown menu out for now.